+++ src/core/types/events.ts
@@ -3,7 +3,7 @@
  * @module core/types/events
  */
 
-import { NostrEventKind, NostrMessageType } from '../constants';
+import { NostrEventKind, NostrMessageType } from '../constants.js';
 
 /**
  * Basic Nostr event structure